+++ core/rend/transform_matrix.h
@@ -87,8 +87,8 @@ public:
 
 		if (renderingContext->isRTT)
 		{
-			dcViewport.x = renderingContext->fb_X_CLIP.max - renderingContext->fb_X_CLIP.min + 1;
-			dcViewport.y = renderingContext->fb_Y_CLIP.max - renderingContext->fb_Y_CLIP.min + 1;
+			dcViewport.x = (float)(renderingContext->fb_X_CLIP.max - renderingContext->fb_X_CLIP.min + 1);
+			dcViewport.y = (float)(renderingContext->fb_Y_CLIP.max - renderingContext->fb_Y_CLIP.min + 1);
 			normalMatrix = glm::translate(glm::vec3(-1, -rttFlipY, 0))
 				* glm::scale(glm::vec3(2.0f / dcViewport.x, 2.0f / dcViewport.y * rttFlipY, 1.f));
 			scissorMatrix = normalMatrix;
